# monopatines/mvc_facturacion/mvc_facturacion_swagger.yaml
openapi: 3.0.3
info:
  title: Microservicio de Facturación
  description: API para gestión de tarifas, suscripciones y cargas del sistema
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Gateway local

tags:
  - name: Tarifas
    description: Operaciones relacionadas con tarifas
  - name: Subscripciones
    description: Operaciones relacionadas con suscripciones
  - name: Cargas
    description: Operaciones relacionadas con cargas

paths:
  /tarifas:
    get:
      tags: ["Tarifas"]
      summary: Listar todas las tarifas
      description: Obtiene la lista completa de tarifas registradas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de tarifas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TarifaResponse'
    post:
      tags: ["Tarifas"]
      summary: Crear una nueva tarifa
      description: Registra una nueva tarifa en el sistema
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TarifaRequest'
      responses:
        '200':
          description: Tarifa creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TarifaResponse'

  /tarifas/{id}:
    get:
      tags: ["Tarifas"]
      summary: Obtener una tarifa por ID
      description: Recupera la información de una tarifa específica
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Tarifa encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TarifaResponse'
    delete:
      tags: ["Tarifas"]
      summary: Eliminar una tarifa
      description: Elimina una tarifa del sistema
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Tarifa eliminada exitosamente
    put:
      tags: ["Tarifas"]
      summary: Actualizar una tarifa existente
      description: Modifica la información de una tarifa
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TarifaRequest'
      responses:
        '200':
          description: Tarifa actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TarifaResponse'

  /tarifas/vigente:
    get:
      tags: ["Tarifas"]
      summary: Obtener tarifa vigente
      description: Recupera la tarifa vigente para una fecha específica o la fecha actual
      security:
        - bearerAuth: []
      parameters:
        - name: fecha
          in: query
          required: false
          schema:
            type: string
            format: date
          example: "2025-11-15"
      responses:
        '200':
          description: Tarifa vigente encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TarifaResponse'

  /subscripciones:
    get:
      tags: ["Subscripciones"]
      summary: Listar todas las suscripciones
      description: Obtiene la lista completa de suscripciones registradas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de suscripciones obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscripcionResponse'
    post:
      tags: ["Subscripciones"]
      summary: Crear una nueva suscripción
      description: Registra una nueva suscripción en el sistema
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscripcionRequest'
      responses:
        '200':
          description: Suscripción creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscripcionResponse'

  /subscripciones/{id}:
    get:
      tags: ["Subscripciones"]
      summary: Obtener una suscripción por ID
      description: Recupera la información de una suscripción específica
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Suscripción encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscripcionResponse'
    delete:
      tags: ["Subscripciones"]
      summary: Eliminar una suscripción
      description: Elimina una suscripción del sistema
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Suscripción eliminada exitosamente
    put:
      tags: ["Subscripciones"]
      summary: Actualizar una suscripción existente
      description: Modifica la información de una suscripción
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscripcionRequest'
      responses:
        '200':
          description: Suscripción actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscripcionResponse'

  /cargas:
    get:
      tags: ["Cargas"]
      summary: Listar todas las cargas
      description: Obtiene la lista completa de cargas registradas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de cargas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CargaResponse'
    post:
      tags: ["Cargas"]
      summary: Crear una nueva carga
      description: Registra una nueva carga en el sistema
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CargaRequest'
      responses:
        '200':
          description: Carga creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CargaResponse'

  /cargas/{id}:
    get:
      tags: ["Cargas"]
      summary: Obtener una carga por ID
      description: Recupera la información de una carga específica
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Carga encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CargaResponse'
    delete:
      tags: ["Cargas"]
      summary: Eliminar una carga
      description: Elimina una carga del sistema
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Carga eliminada exitosamente
    put:
      tags: ["Cargas"]
      summary: Actualizar una carga existente
      description: Modifica la información de una carga
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CargaRequest'
      responses:
        '200':
          description: Carga actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CargaResponse'

components:
  schemas:
    TarifaRequest:
      type: object
      required:
        - precioPorMinuto
        - precioPorMinutoExtendido
        - mensualidadPremium
        - vigencia
      properties:
        precioPorMinuto:
          type: number
          format: double
          example: 5.50
        precioPorMinutoExtendido:
          type: number
          format: double
          example: 3.00
        mensualidadPremium:
          type: number
          format: double
          example: 1500.00
        vigencia:
          $ref: '#/components/schemas/Vigencia'

    TarifaResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        precioPorMinuto:
          type: number
          format: double
          example: 5.50
        precioPorMinutoExtendido:
          type: number
          format: double
          example: 3.00
        mensualidadPremium:
          type: number
          format: double
          example: 1500.00
        vigencia:
          $ref: '#/components/schemas/Vigencia'

    SubscripcionRequest:
      type: object
      required:
        - idCuenta
        - periodoFacturacion
        - fechaPago
        - monto
        - vigencia
      properties:
        idCuenta:
          type: integer
          example: 1001
        periodoFacturacion:
          type: string
          format: year-month
          example: "2025-11"
        fechaPago:
          type: string
          format: date
          example: "2025-11-15"
        monto:
          type: integer
          example: 1500
        estado:
          $ref: '#/components/schemas/EstadoSubscripcion'
        vigencia:
          $ref: '#/components/schemas/Vigencia'

    SubscripcionResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        idCuenta:
          type: integer
          example: 1001
        periodoFacturacion:
          type: string
          format: year-month
          example: "2025-11"
        fechaPago:
          type: string
          format: date
          example: "2025-11-15"
        monto:
          type: integer
          example: 1500
        estado:
          $ref: '#/components/schemas/EstadoSubscripcion'
        vigencia:
          $ref: '#/components/schemas/Vigencia'

    CargaRequest:
      type: object
      required:
        - viajeId
        - cuentaId
      properties:
        viajeId:
          type: integer
          example: 5001
        cuentaId:
          type: integer
          example: 1001
        montoTotal:
          type: integer
          example: 2500
        cargaNormal:
          type: integer
          example: 2000
        cargaPausaExtendida:
          type: integer
          example: 500
        duracionMinutos:
          type: integer
          example: 45
        duracionPausaMinutos:
          type: integer
          example: 15
        kmRecorridos:
          type: number
          format: double
          example: 8.5

    CargaResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        viajeId:
          type: integer
          example: 5001
        cuentaId:
          type: integer
          example: 1001
        montoTotal:
          type: integer
          example: 2500
        cargaNormal:
          type: integer
          example: 2000
        cargaPausaExtendida:
          type: integer
          example: 500
        duracionMinutos:
          type: integer
          example: 45
        duracionPausaMinutos:
          type: integer
          example: 15
        kmRecorridos:
          type: number
          format: double
          example: 8.5
        fechaCarga:
          type: string
          format: date-time
          example: "2025-11-15T10:30:00"

    Vigencia:
      type: object
      required:
        - inicio
      properties:
        inicio:
          type: string
          format: date
          example: "2025-01-01"
        fin:
          type: string
          format: date
          nullable: true
          example: "2025-12-31"

    EstadoSubscripcion:
      type: string
      enum: [ACTIVA, INACTIVA, SUSPENDIDA]
      example: "ACTIVA"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT